pyproject.toml
[project]
name = "simple-wikipedia-dataset"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10.6"

dependencies = [
    "accelerate>=1.2.1",
    "datasets>=3.2.0",
    "fastparquet>=2024.11.0",
    "html2text>=2024.2.26",
    "jupyter>=1.1.1",
    "langchain-community>=0.3.13",
    "langchain>=0.3.13",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "torch>=2.5.1",
    "tqdm>=4.67.1",
    "transformers>=4.47.1",
    "wikitextparser>=0.56.3",
    "openai>=1.58.1",
    "langchain-openai>=0.2.14",
    "matplotlib>=3.10.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
]

[tool.ruff]
line-length = 88  # black default
indent-width = 4  # black default

[tool.ruff.format]
# Enable reformatting of code snippets in docstrings.
docstring-code-format = true

[tool.ruff.lint]
select = [
    "B",  # flake8-bugbear
    "C",  # flake8-comprehensions
    "D",  # pydocstyle
    "D213",  # docstring starts 2nd line
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "W",  # pycodestyle warnings
]
ignore = [
    "D212",  # incompatible with D213
]
task-tags = ["TODO", "todo"]

[tool.ruff.lint.pycodestyle]
ignore-overlong-task-comments = true

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
ignore_missing_imports = true
warn_no_return = false
disallow_untyped_defs = false
allow_redefinition = true

[tool.isort]
profile = "black"
force_sort_within_sections = true
sections = "STDLIB,THIRDPARTY,LOCALFOLDER"
import_headings = "standard library,third-party imports,local imports"
order_by_type = true
force_single_line = false
combine_as_imports = true
line_length = 88
src_paths = ["src"]
